+++ package.json
@@ -13,20 +13,24 @@
     "bignumber.js": "^9.0.1",
     "bn.js": "^5.1.3",
     "decimal.js": "^10.2.1",
-    "decimal.js-light": "^2.5.1",
-    "ethers": "^5.0.29"
+    "decimal.js-light": "^2.5.1"
   },
   "devDependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.2",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
-    "hardhat": "^2.1.1"
+    "assert": "^2.0.0",
+    "chai": "^4.3.4",
+    "ethers": "^5.0.29",
+    "hardhat": "^2.1.1",
+    "hardhat-deploy": "^0.7.0-beta.50",
+    "hardhat-deploy-ethers": "^0.3.0-beta.7"
   },
   "scripts": {
     "run-node": "hardhat node",
     "compile": "hardhat compile",
     "deploy:local": "hardhat run ./scripts/deployContracts.js --network localhost",
     "deploy:mainnet": "hardhat run ./scripts/deployContracts.js --network mainnet",
-    "test": "hardhat test --network localhost"
+    "test": "hardhat test --network hardhat"
   },
   "repository": {
     "type": "git",
